add memcached_controller_test
