+++ controllers/memcached_controller_test.go
@@ -1,98 +1,3 @@
-# 7. Write controller tests
-
-## Tools
-
-1. [envtest](https://pkg.go.dev/sigs.k8s.io/controller-runtime/pkg/envtest) provides libraries for integration testing by starting a local control plane. (`etcd` an `kube-apiserver`)
-1. [Ginkgo](https://pkg.go.dev/github.com/onsi/ginkgo) BDD framework.
-1. [Gomega](https://pkg.go.dev/github.com/onsi/gomega) Matcher library for testing.
-
-## Prepare `suite_test.go`
-
-1. Import necessary packages.
-    ```diff
-     import (
-    +       "context"
-            "path/filepath"
-            "testing"
-    +       ctrl "sigs.k8s.io/controller-runtime"
-    +
-            . "github.com/onsi/ginkgo"
-            . "github.com/onsi/gomega"
-            "k8s.io/client-go/kubernetes/scheme"
-    -       "k8s.io/client-go/rest"
-            "sigs.k8s.io/controller-runtime/pkg/client"
-            "sigs.k8s.io/controller-runtime/pkg/envtest"
-            "sigs.k8s.io/controller-runtime/pkg/envtest/ter"
-            logf "sigs.k8s.io/controller-runtime/pkg/log"
-            "sigs.k8s.io/controller-runtime/pkg/log/zap"
-    +       "sigs.k8s.io/controller-runtime/pkg/manager"
-    ```
-
-1. Prepare global variables.
-    ```diff
-    -var cfg *rest.Config
-    -var k8sClient client.Client
-    -var testEnv *envtest.Environment
-    +var (
-    +       k8sClient  client.Client
-    +       k8sManager manager.Manager
-    +       testEnv    *envtest.Environment
-    +       ctx        context.Context
-    +       cancel     context.CancelFunc
-    +)
-    ```
-
-1. Add the following lines at the end of `BeforeSuite` in `controllers/suite_test.go`.
-
-    ```go
-        // Create context with cancel.
-        ctx, cancel = context.WithCancel(context.TODO())
-
-        // Register the schema to manager.
-        k8sManager, err = ctrl.NewManager(cfg, ctrl.Options{
-            Scheme: scheme.Scheme,
-        })
-
-        // Initialize `MemcachedReconciler` with the manager client schema.
-        err = (&MemcachedReconciler{
-            Client: k8sManager.GetClient(),
-            Scheme: k8sManager.GetScheme(),
-        }).SetupWithManager(k8sManager)
-
-        // Start the with a goroutine.
-        go func() {
-            defer GinkgoRecover()
-            err = k8sManager.Start(ctx)
-            Expect(err).ToNot(HaveOccurred(), "failed to run ger")
-        }()
-    ```
-
-1. Add `cancel()` to AfterSuite.
-
-    ```diff
-     var _ = AfterSuite(func() {
-    +       cancel()
-            By("tearing down the test environment")
-            err := testEnv.Stop()
-            Expect(err).NotTo(HaveOccurred())
-    ```
-
-## Write controller's tests
-
-Test cases in `controllers/memcached_controller_test.go`:
-1. When `Memcached` is created
-    1. `Deployment` should be created.
-    1. `Memcached`'s nodes have pods' names.
-1. When `Memcached`'s `size` is updated
-    1. `Deployment`'s `replicas` should be updated.
-    1. `Memcached`'s nodes have new pods' names.
-1. When `Deployment` is updated
-    1. Deleting `Deployment` -> `Deployment` is recreated.
-    1. Updating `Deployment` with `replicas = 0` -> `Deployment`'s replicas is updated to the original number.
-
-<details><summary>memcached_controller_test.go</summary>
-
-```go
 package controllers
 
 import (
@@ -323,28 +228,3 @@ func newPod(name string) *v1.Pod {
 		Status: v1.PodStatus{},
 	}
 }
-```
-
-</details>
-
-## Run the tests
-
-```
-make test
-```
-
-<details>
-
-```
-/Users/nakamasato/repos/nakamasato/memcached-operator/bin/controller-gen rbac:roleName=manager-role crd webhook paths="./..." output:crd:artifacts:config=config/crd/bases
-/Users/nakamasato/repos/nakamasato/memcached-operator/bin/controller-gen object:headerFile="hack/boilerplate.go.txt" paths="./..."
-go fmt ./...
-go vet ./...
-GOBIN=/Users/nakamasato/repos/nakamasato/memcached-operator/bin go install sigs.k8s.io/controller-runtime/tools/setup-envtest@latest
-KUBEBUILDER_ASSETS="/Users/nakamasato/Library/Application Support/io.kubebuilder.envtest/k8s/1.23.3-darwin-amd64" go test ./... -coverprofile cover.out
-?       github.com/example/memcached-operator   [no test files]
-?       github.com/example/memcached-operator/api/v1alpha1      [no test files]
-ok      github.com/example/memcached-operator/controllers       18.284s coverage: 79.3% of statements
-```
-
-</details>
